third_party/WebKit/LayoutTests/inspector/tracing/timeline-style-recalc-with-invalidations-expected.txt

Tests the Timeline API instrumentation of style recalc events with invalidations.

PASS
PASS
PASS


Running: testLocalFrame

Running: multipleStyleRecalcs

Running: testSubframe
